requirements.txt
streamlit==1.22.0

# LangChain core + community modules
langchain==0.0.284
langchain-community==0.0.19
langchain-google-genai==0.0.6

# Google Gemini SDK
google-generativeai==0.4.1

# Embeddings + FAISS vector store
faiss-cpu==1.7.4

# Web document loader + parsing dependencies
unstructured==0.9.2
python-magic==0.4.27
python-magic-bin==0.4.14
libmagic==1.0

# Environment config
python-dotenv==1.0.0

# Tokenization support
tiktoken==0.4.0
